requirements.txt
python-dotenv==1.2.2
discord.py==2.7.1